Update the classname with cn and rename the constants folder
